Modifications by Ivailo Monev I've modified the hook and the buildiso script so that file .miso is created so that the miso hook can make use of it and on boot list all removable devices found via /sys/block/*/removable, mount it and if /bootmnt/.miso exists the boot continues as usual.
The use case (as I've mentioned in one of my emails to the manjaro-dev mailing list) is to get rid of the misolabel restriction where the label of the boot device should match the one set at boot or the kernel paramater should be changed. Now this is annoying if you are running Windows and use unetbootin or universal-usb-installer to put the ISO image to USB-pendrive and you don't change the label, the boot will fail. But the problem can accure even if Linux is used but the image is not dumped via dd. Another thing is that you can't have a pendrive with multiple Manjaro images on it or you must change the label kernel paramaters of all of them to be the exact same one.

The draw back is only that poll_device() is not used which means that the hook won't wait 30-seconds for the boot device to show up (which may lead to another problem with delayed udev tirgger for USB devices) and for this 'miso_wait' is introduced so that the hook will wait the value of the argument (i.e. miso_wait=10) in seconds before looking for a boot device.